backend/requirements.txt
Django==2.2.16
djangorestframework==3.12.4
django-filter==2.4.0

# asgiref==3.2.10
# drf_extra_fields==3.4.0
# gunicorn==20.0.4
psycopg2-binary==2.8.6
# python-dotenv==0.20.0
# pytz==2020.1
# requests==2.26.0
# sqlparse==0.3.1

# Не встают на место, без шаманизма.
# Нужно обновить pip install --upgrade pip
# django-colorfield==0.7.2
# djoser==2.1.0
# Pillow==9.2.0
